[circt-reduce] Move TieOffCache to FIRRTLUtils
Move TieOffCache helper class to FIRRTLUtils.h and FIRRTLUtils.cpp
to avoid duplication across multiple files.

AI-assisted-by: Augment (Claude Sonnet 4.5)
