+++ lib/Dialect/FIRRTL/FIRRTLUtils.cpp
@@ -23,6 +23,28 @@
 using namespace circt;
 using namespace firrtl;
 
+//===----------------------------------------------------------------------===//
+// TieOffCache
+//===----------------------------------------------------------------------===//
+
+Value TieOffCache::getInvalid(FIRRTLBaseType type) {
+  Value &cached = cache[type];
+  if (!cached)
+    cached = InvalidValueOp::create(builder, type);
+  return cached;
+}
+
+Value TieOffCache::getUnknown(PropertyType type) {
+  Value &cached = cache[type];
+  if (!cached)
+    cached = builder.create<UnknownValueOp>(type);
+  return cached;
+}
+
+//===----------------------------------------------------------------------===//
+// emitConnect
+//===----------------------------------------------------------------------===//
+
 void circt::firrtl::emitConnect(OpBuilder &builder, Location loc, Value dst,
                                 Value src) {
   ImplicitLocOpBuilder locBuilder(loc, builder.getInsertionBlock(),
